Rename container names. Now container name convention changes, higher order components related to container moved from root folder to the same location level as container itself
