+++ src/app/containers/ConfigContainer/configHoc.js
