+++ src/app/containers/ConfigNavigationContainer/configNavigationHoc.js
