+++ src/app/containers/ConfigPageCategoriesContainer/configPageCategoriesHoc.js
