+++ src/app/containers/ConfigProductCategoriesContainer/configProductCategoriesHoc.js
