+++ src/app/containers/PagesContainer/pagesHoc.js
