+++ src/app/containers/ProductViewContainer/mocks/checkout.js
