+++ src/app/containers/ProductViewContainer/mocks/config.js
