+++ src/app/containers/ProductViewContainer/mocks/products.js
