+++ src/app/containers/ProductViewContainer/productViewHoc.js
