+++ src/app/containers/ProductsContainer/productsHoc.js
